refactor: Move the ffi part of the verifier into the ffi folder, and move the internals down a level
